Extract dump information to separate file
